% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMax.R
\name{colMax}
\alias{colMax}
\title{Max of each column}
\usage{
colMax(Matrix)
}
\arguments{
\item{Matrix}{A \code{matrix} object or a \code{data-frame}}
}
\value{
Returns a vector of \code{length = ncol(Matrix)} with the column maximums
}
\description{
Get the Max of each column of a matrix or data-frame
}
\details{
The values should be all numeric
}
\examples{
data( LifeCycleSavings )
colMax( Matrix = t(LifeCycleSavings) )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{rowMax}
}
\keyword{maximum}

