% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMin.R
\name{colMin}
\alias{colMin}
\title{Min of each column}
\usage{
colMin(Matrix)
}
\arguments{
\item{Matrix}{A \code{matrix} object or a \code{data-frame}}
}
\value{
Returns a vector of \code{length = ncol(Matrix)} with the row Minimums
}
\description{
Get the Min of each column of a matrix or data-frame
}
\details{
The values should be all numeric
}
\examples{
data( LifeCycleSavings )
colMin( Matrix = t( LifeCycleSavings ) )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{rowMin}
}
\keyword{minimum}

