% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colSD.R
\name{colSD}
\alias{colSD}
\title{Standard deviation of each column}
\usage{
colSD(Matrix)
}
\arguments{
\item{Matrix}{A \code{matrix} object or a \code{data-frame}}
}
\value{
Returns a vector of \code{length = ncol(Matrix)} with the column SDs
}
\description{
Gets the standard deviation (SD) of each column of a matrix or data-frame
}
\details{
The values should be all numeric
}
\examples{
data( LifeCycleSavings )
colSD( Matrix = t( LifeCycleSavings ) )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{rowSD}
}
\keyword{standard-deviation}

