% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colSE.R
\name{colSE}
\alias{colSE}
\title{Standard Error of mean of columns}
\usage{
colSE(Matrix)
}
\arguments{
\item{Matrix}{A \code{matrix} object or a \code{data-frame}}
}
\value{
Returns a vector of \code{length = ncol(Matrix)} with the column SDs
}
\description{
Returns the standard error (SE) of mean of columns a matrix or data frame
}
\details{
The values should be all numeric
}
\examples{
data( LifeCycleSavings )
colSE( Matrix = t( LifeCycleSavings ) )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{rowSE}, \code{colSD}
}
\keyword{standard-error}

