% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colVar.R
\name{colVar}
\alias{colVar}
\title{Variance of each column}
\usage{
colVar(Matrix)
}
\arguments{
\item{Matrix}{A \code{matrix} object or a \code{data-frame}}
}
\value{
Returns a vector of \code{length = ncol(Matrix)} with the column variances
}
\description{
Returns the Variance of each column of a dataset
}
\details{
The values should be all numeric
}
\examples{
data( LifeCycleSavings )
colVar( Matrix = t( LifeCycleSavings ) )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{rowVar}
}
\keyword{variance}

