% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pValue_lm.R
\name{get_pValue_lm}
\alias{get_pValue_lm}
\title{Get p-value from \code{lm}}
\usage{
get_pValue_lm(lmobject, Round = 6)
}
\arguments{
\item{lmobject}{This should be an output from \code{lm}}

\item{Round}{The desired number of places of decimals for the p-value}
}
\value{
pValue of the whole model is returned
}
\description{
Returns the model p-value, rounded to desired number of decimals, from a \code{lm} object
}
\examples{
data( cars )
get_pValue_lm( lm( speed ~ dist, data = cars ), 4 )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{lm}
}
\keyword{p-value}

