% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last}
\alias{last}
\title{Last element of a vector or list}
\usage{
last(x)
}
\arguments{
\item{x}{A vector or list}
}
\value{
The last element of the vector or list
}
\description{
Returns the last element of a vector or list
}
\examples{
vector <- c( 2, 3, 5, 2, 1 )
last( vector )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{tail}
}
\keyword{last}
\keyword{tail}

