% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Mode of a vector}
\usage{
mode(x)
}
\arguments{
\item{x}{A vector of numeric of character values}
}
\value{
Returns the mode. It may or may not be unique.
}
\description{
Returns the Mode (or most frequent element) of a vector of numeric of character values
}
\examples{
x = c( 1, 1, 4, 5, 6, 6 )
mode( x )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{lm}
}
\keyword{frequency}
\keyword{mode}

