% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.is.x.R
\name{na.is.x}
\alias{na.is.x}
\title{Convert NAs to "x" in a vector}
\usage{
na.is.x(X, replaceWith = "x")
}
\arguments{
\item{X}{Vector with NAs in it}

\item{replaceWith}{Value by which NAs should be replaced}
}
\value{
Returns the same vector with NA's replaced
}
\description{
Convert NAs to user defined value in a vector
}
\examples{
vector <- c(2, 3, 5, NA, 1)
na.is.x( X = vector, replaceWith = 100 )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{na.omit}, \code{na.action}, \code{na.is.zero}
}

