% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.is.zero.R
\name{na.is.zero}
\alias{na.is.zero}
\title{Convert NAs to Zero in a vector}
\usage{
na.is.zero(X)
}
\arguments{
\item{x}{Vector with NAs in it}
}
\value{
Returns the same vector with zero's instead of NA's
}
\description{
Convert NAs to Zero in a vector
}
\examples{
vector <- c(2, 3, 5, NA, 1)
na.is.zero(vector)

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{na.omit}, \code{na.action}
}

