% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_vector.R
\name{normalize_vector}
\alias{normalize_vector}
\title{Normalize a vector}
\usage{
normalize_vector(Vector, withSD = TRUE)
}
\arguments{
\item{Vector}{A vector of numerical values}

\item{withSD}{A Boolean indicating if divide by SD when normalizing. Defaults to \code{TRUE}}
}
\value{
Returns a vector of same length, with normalized values
}
\description{
Normalize a vector by subtracting its mean from all the elements and dividing by its SD
}
\examples{
vector <- c( 2, 3, 5, 2, 1 )
normalize_vector( vector, TRUE )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\keyword{normalize}

