% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeMatrixWithSD.R
\name{normalizeMatrixWithRowSD}
\alias{normalizeMatrixWithRowSD}
\title{Normalize columns of a matrix by rowSD}
\usage{
normalizeMatrixWithRowSD(matrixData)
}
\arguments{
\item{matrixData}{A matrix of numerical values}
}
\value{
Returns a matrix of same dimensions, with normalized values
}
\description{
Normalize columns of a matrix by dividing each row by rowSD.
}
\details{
The resulting normalized matrix will have rowSD = 1 for each row.
}
\examples{
data( LifeCycleSavings )
LifeCycleSavingsNorm <- normalizeMatrixWithRowSD( matrixData = t( LifeCycleSavings ) )
rowSD( LifeCycleSavingsNorm )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{normalizeMatrix}, \code{rowSD}
}
\keyword{normalize}

