% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Opposite of \code{\%in\%}}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{Element to check if it is in y}

\item{y}{Element to check if x is in}
}
\value{
Returns the elements of x not in y
}
\description{
Opposite of \code{\%in\%}: If an element x is not in y, then returns TRUE
}
\details{
name notin
}
\examples{
x <- 3
y <- c(4, 1, 5, 6)
x \\\%notin\\\% y
}
\author{
Subhrangshu Nandi; \email{nands31@gmail.com}
}
\keyword{in}

