% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctChange.R
\name{pctChange}
\alias{pctChange}
\title{Returns percentage change between two numbers}
\usage{
pctChange(Before, After)
}
\arguments{
\item{Before}{A numeric variable}

\item{After}{A numeric variable}
}
\value{
PctChange A numeric variable in percentage
}
\description{
Returns percentage change between two numbers
}
\examples{
pctChange( Before = 10, After = 11 )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}

