% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMax.R
\name{rowMax}
\alias{rowMax}
\title{Max of each row}
\usage{
rowMax(Matrix)
}
\arguments{
\item{Matrix}{A \code{matrix} object or a \code{data-frame}}
}
\value{
Returns a vector of \code{length = nrow(Matrix)} with the column maximums
}
\description{
Get the Max of each row of a matrix or data-frame
}
\details{
The values should be all numeric
}
\examples{
data( LifeCycleSavings )
rowMax( Matrix = t(LifeCycleSavings) )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{colMax}
}
\keyword{maximum}

