% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowSD.R
\name{rowSD}
\alias{rowSD}
\title{Standard deviation of each row}
\usage{
rowSD(Matrix)
}
\arguments{
\item{Matrix}{A \code{matrix} object or a \code{data-frame}}
}
\value{
Returns a vector of \code{length = nrow(Matrix)} with the row SDs
}
\description{
Gets the standard deviation (SD) of each row of a matrix or data-frame
}
\details{
The values should be all numeric
}
\examples{
data( LifeCycleSavings )
rowSD( Matrix = t( LifeCycleSavings ) )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{colSD}
}
\keyword{standard-deviation}

