% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowSE.R
\name{rowSE}
\alias{rowSE}
\title{Standard Error of mean of rows}
\usage{
rowSE(Matrix)
}
\arguments{
\item{Matrix}{A \code{matrix} object or a \code{data-frame}}
}
\value{
Returns a vector of \code{length = nrow(Matrix)} with the row SDs
}
\description{
Returns the standard error (SE) of mean of rows a matrix or data frame
}
\details{
The values should be all numeric
}
\examples{
data( LifeCycleSavings )
rowSE( Matrix = t( LifeCycleSavings ) )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{colSE}, \code{colSD}
}
\keyword{standard-error}

