% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowVar.R
\name{rowVar}
\alias{rowVar}
\title{Variance of each row}
\usage{
rowVar(Matrix)
}
\arguments{
\item{Matrix}{A \code{matrix} object or a \code{data-frame}}
}
\value{
Returns a vector of \code{length = nrow(Matrix)} with the row variances
}
\description{
Returns the Variance of each row of a dataset
}
\details{
The values should be all numeric
}
\examples{
data( LifeCycleSavings )
rowVar( Matrix = t( LifeCycleSavings ) )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{colVar}
}
\keyword{variance}

