% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRows.R
\name{sampleRows}
\alias{sampleRows}
\title{Returns a subset of data-frame or matrix}
\usage{
sampleRows(origData, Seed = 1, sampleSize, replaceBoolean = FALSE)
}
\arguments{
\item{origData}{A \code{matrix} object or a \code{data-frame}}

\item{Seed}{To randomize, specify any Seed you desire}

\item{sampleSize}{The number of rows to sample, from the original dataset}

\item{replaceBoolean}{Whether to sample with replacement or not. Default is \code{FALSE}}
}
\value{
Returns an object of same type as original dataset, with sampled columns
}
\description{
Returns a random subset of rows of a data-frame or matrix
}
\examples{
data( LifeCycleSavings )
sampleRows( LifeCycleSavings, Seed = 1, sampleSize = 10, FALSE )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\seealso{
\code{sample}, \code{subset}, \code{sampleColumns}
}
\keyword{random}
\keyword{sample}

