% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_vector.R
\name{truncate_vector}
\alias{truncate_vector}
\title{Truncate a vector on one/both sides}
\usage{
truncate_vector(Vector, pct_trunc_left = 10, pct_trunc_right = 10)
}
\arguments{
\item{Vector}{A vector of numerical values}

\item{pct_trunc_left}{A number denoting the percentage to truncate from left. For 10\% enter 10.}

\item{pct_trunc_right}{A number denoting the percentage to truncate from right. For 10\% enter 10.}
}
\value{
Returns a truncated vector
}
\description{
Truncate (or trim) a vector on one or both sides, by user defined percentages
}
\examples{
vector <- c(2, 3, 5, 2, 1, 2, 3, 5, 2, 1, 2, 3, 5, 2, 1, 2, 3, 5, 2, 1, 2, 3, 5, 2, 1)
truncate_vector(vector, 5, 5)

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\keyword{trim}
\keyword{truncate}

