% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/without.R
\name{without}
\alias{\%w/o\%}
\alias{without}
\title{Return the elements of x not in y}
\usage{
x \%w/o\% y
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\value{
Returns the elements of x not in y
}
\description{
Return the elements of x not in y. In set notation, this returns A \\ B
}
\examples{
x <- c(4, 1, 5, 6) 
y <- c(4, 5) 
x \\\%w/o\\\% y

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\keyword{without}

