% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/without.R
\name{\%w/o\%}
\alias{\%w/o\%}
\title{Return the elements of x not in y}
\usage{
x \%w/o\% y
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\value{
Returns the elements of x not in y
}
\description{
Return the elements of x not in y. In set notation, this returns A \\ B
}
\details{
name without
}
\examples{
x <- c(4, 1, 5, 6) 
y <- c(4, 5) 
x \\\%w/o\\\% y
}
\author{
Subhrangshu Nandi; \email{nands31@gmail.com}
}
\keyword{without}

